Expand fuzzy  match module to include array matcher
